eeprom: add eeprom_region struct
** Cherry-picked from the master branch. **

This adds a struct which can be used to specify region information
known to exist on a particular platform. It will be used for
programatically finding regions of a given type.

BUG=chrome-os-partner:16132
BRANCH=spring
TEST=Used this structure members in a later change (also by dhendrix) to read
the event log on spring.

Change-Id: Iff098b739f5a4ec822a199db77a4a27fac0eb35d
Reviewed-on: https://gerrit.chromium.org/gerrit/50021
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58512
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>