UPSTREAM: kconfig/lxdialog: get ncurses CFLAGS with pkg-config
This makes "make menuconfig" also work on systems where ncurses is not
installed in a standard location (such as on NixOS).

This patch changes ccflags() so that it tries pkg-config first, and only
if pkg-config fails does it go back to the fallback/manual checks. This
is the same algorithm that ldflags() already uses.

BUG=None
TEST=`make menuconfig` works again
BRANCH=None

Change-Id: Ie5e0c5d77585c4ebcb0d831b808b2f72fd8757ca
Signed-off-by: Bjrn Forsman <bjorn.forsman@gmail.com>
Signed-off-by: Michal Marek <mmarek@suse.cz>
Reviewed-on: https://chromium-review.googlesource.com/685963
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Shawn N <shawnn@chromium.org>