veyron: Re-factor probe function
Veyron derivatives are becoming rather diverse. This updates the probe
function to allow for platform attributes to be specified in the
list of veyron variants along with the FDT name.

One of the benefits is that we have a central location for FDT name
and canonical platform name so we no longer need the series of
if-statements in the probe function.

The follow-up patch will add non-laptop devices to the list, so the
has_ec flag will be set to 0 and the call to pinky_ec_setup() will
be skipped in the post-setup phase.

BUG=none
BRANCH=none
TEST=built and run on Brain w/ follow-up CL

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: I2657b22f49c84f198ebe76c053967cc973730be5
Reviewed-on: https://chromium-review.googlesource.com/245520
Reviewed-by: Shawn N <shawnn@chromium.org>
(cherry picked from commit d6322adbc0ed1aedb7cbd24a166ccd82a2c2065e)
Reviewed-on: https://chromium-review.googlesource.com/246067
Commit-Queue: Shunqian Zheng <zhengsq@rock-chips.com>
Tested-by: Shunqian Zheng <zhengsq@rock-chips.com>