flashrom: Find programs in default PATH instead of using `which`.
The utility 'which' is not a shell built-in command, and what it does
is simpy trying to find programs in default PATH. This implies calling
`which` then use the output for complete path sent to execv is
identical to directly calling execvp (which uses PATH).

In issue chromium:763264, we think in special environment like
initramfs, calling 'which' may not work properly even if PATH is set.

This change tries to replace execv by execvp, and simply use 'flashrom'
if the system is not Android; which reduced dependency and complexity of
calling flashrom.

BUG=chromium:763264
TEST=emerge-reef mosys; mosys nvram vboot read

Change-Id: Iea1e2dacbdf59b2f14a25b681c7e5b0001b046ca
Reviewed-on: https://chromium-review.googlesource.com/658578
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@google.com>
Reviewed-by: David Hendricks <david.hendricks@gmail.com>