Obtain the coral SKU ID table from the master configuration
At present mosys has its own table which maps from SKU ID to model and
brand code. This duplicates information in the master configuration and
goes against a design goal of unified builds, which is to unify the
configuration.

Embed the master configuration into mosys and use that to report the model
and brand code. This means that the SKU mapping is maintained in the same
place as other configuration.

With this extra code and the master configuration itself, the statically
linked mosys_s increases in size by 11400 bytes to 1344736 bytes on coral
at present, which seems acceptable. The code size increment is
approximately 9KB.

BUG=chromium:770183
BRANCH=none
TEST=with other changes: FEATURES=test emerge-coral --nodeps mosys
Also manual test on robo

Change-Id: Ia9934dd1a8707fd75a20fcbd55d850e9dc0aef1c
Reviewed-on: https://chromium-review.googlesource.com/692478
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>