Add support for accessing the SKU ID map and brand code
These are the only two pieces of information that mosys currently needs
from the master configuration. Most likely we will use cros_config for
other things.

Add support for reading from an embedded device tree. This will be used to
replace the internal table. The device tree is called config.dtb and is
written to lib/cros_config/ by the ebuild just before the mosys build
starts.

Control this with a new CONFIG_CROS_CONFIG which is off by default. Boards
which use unified builds can enable it (in the ebuild).

BUG=chromium:770183
BRANCH=none
TEST=with other changes: FEATURES=test emerge-coral --nodeps mosys

Change-Id: Ic2a70a566e282ac1d96f05cd51917c995265e6d7
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/692476
Reviewed-by: Jason Clinton <jclinton@chromium.org>