Add tests for cros_config
At present the mosys tests are not ideal. They don't run from the ebuild
and they require cmockery, which does not appear to be available in our
portage. Also there appear to be build errors (discovered after an attempt
to run the tests). It looks like the code that is being tested does not
change much, so perhaps the tests were determined to have little value,
although it seems that quite a bit of effort was put in at some point.

Add a new simple test program, which does not need cmockery, with a single
test of the cros_config library. Include a suitable test configuration
(taken from coral) which exercises the features.

BUG=chromium:770183
BRANCH=none
TEST=with other changes: FEATURES=test emerge-coral --nodeps mosys

Change-Id: I72b883b6836aab80d7fa69f760ffe4078ff71ad2
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/692477
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: C Shapiro <shapiroc@google.com>