Allow cros_config to read the SKU ID
At present the caller passes in the SKU ID. Once we start looking up the
SMBIOS name this will not make sense, since the binding which describes
how to do the lookup is managed by cros_config. The binding currently
says it uses the SMBIOS name and SKU ID, but we may in future adjust it
to support additional methods (e.g. to support ARM machines).

Adjust the API so that cros_config reads whatever it needs to be able to
figure out the model. That will avoid adding more and more things to the
API call, and giving the fall impression that the caller is in charge of
what is needed.

BUG=chromium:770183
BRANCH=none
TEST=FEATURES=test emerge-coral --nodeps mosys

Change-Id: Ibcca577e070e24640a236b9c9260713f4cce2342
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/704097
Reviewed-by: Jason Clinton <jclinton@chromium.org>