Add a new 'mosys platform id' command
At present with unified builds, mosys identifies the model itself and can
print this information. However, we really would prefer to have mosys
restrict itself to just reading the hardware information, with cros_config
in charge of determining the model, submodel. Otherwise we end up with
mosys duplicating this logic, leading to a desire to share code between
mosys and cros_config.

It is already possible to individually print things that might be related
to platform identity with multiple calls to mosys. But this is
inefficient and also it requires the caller to know which things are
relevant to determining the platform identify.

Add a new 'mosys platform id' command which prints all information
needed to determine the identify, one item per line. Callers should use
-k or -l so they know which value is which.

BUG=chromium:777725
BRANCH=none
TEST=manual test on robo
(mosys has very few tests)

Change-Id: Id69e0627bdbf7e8aece2bedb127760928b91109c
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/737802
Reviewed-by: Jason Clinton <jclinton@chromium.org>