Add support for alternative whitelabel schema
This request came up during a code review for the cros_config support
for the alternative whitelabel schema. At present the firmware updater
calls 'vpd' directly to read the customization ID and uses that as the
signature ID for accessing vblock/rootkey files. With this change it will
be able to call mosys, which in turn calls vpd_get_value which reads the
customization_id from /sys/firmware/vpd/ro/customization_id

The other option here would be to call 'vpd'. In the case where the RO vpd
has been updated since boot, this might be a better option. The downside
is that it takes about 700ms to run. But since this will only be used in
the firmware updater, that doesn't matter.

I've gone with vpd_get_value since that is what mosys uses elsewhere to
read vpd, so I though it best to keep things consistent.

This adds some more whitelabel support to mosys. Hopefully we can drop
this eventually by moving things to use cros_config. For now we need to
implement part of the whitelabel support in mosys to make everything work.
The discussion of moving to using cros_config in the firmware updater is
best had after we have better library support. [1]

With this feature, mosys can obtain the brand code and the signature ID
for whitelabels.

[1] https://chromium-review.googlesource.com/c/chromiumos/platform/firmware/+/668321

BUG=chromium:781928
BRANCH=none
TEST=FEATURES=test emerge-coral mosys
Basic whitelabel santiy test on robo (with config hacked to make robo a
whitelabel)
$ /tmp/mosys platform signature
whitetip1

Change-Id: I40d2db63de66cebbc88c8af34376da70b9a10dc2
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/770653
Reviewed-by: Jason Clinton <jclinton@chromium.org>