mosys: Add a new command to look up the signature ID
Currently in the unibuild firmware updater we call 'vpd' to get access to
the signature ID. A request was made to use mosys instead.

Add a 'mosys platform signature' command to access this. The signature is
will be set by the cros_config identity lookup (in the next CL).

BUG=chromium:775543
BRANCH=none
TEST=manual on robo

Change-Id: Ie70c2cd4e0cfc3e4e3b60292ef01c4d6f47f5fdf
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/770652
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/801478
Commit-Queue: Marco Chen <marcochen@chromium.org>
Tested-by: Marco Chen <marcochen@chromium.org>