Avoid a crash on platforms with no SKU information
At present 'mosys platform signature' assumes that SKU information is
present and crashes if it is not. This command is normally used only with
unified builds, but there is no guarantee that the board actually uses
cro_config to read the SKU information.

Fix this by adding a NULL check.

BUG=chromium:787879
BRANCH=none
TEST=manual test on kahlee

Change-Id: Ifdb351b0dfeaafecd810153d5c851d44edcd514b
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/786133