kahlee: Use cros_config to obtain SKU information
Kahlee uses unified builds so we should switch it to use cros_config to
obtain its SKU information.

To make this work we need to make sure that kahlee support is only built
in if we have CONFIG_CROS_CONFIG defined.

Unfortunately mosys uses an old version of kconfig so it does not seem
possible to do what we do with the kernel (make oldnconfig). I'm not sure
how to work around that other than by adjusting the Makefile.

BUG=chromium:787879
BRANCH=none
TEST=manual test on kahlee

Change-Id: Iecda9568c690f42602dd5fd2286a5db2b6cf1498
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/786134
Reviewed-by: Jason Clinton <jclinton@chromium.org>