Makefile: Handle empty string environment variables
I spent hours looking at this to close an ancient bug
https://github.com/dhendrix/mosys/issues/3 opened in 2012. The reason
that this has fought being resolved for so long is that we are mixing
Makefile and Kbuild systems and somewhere something is setting
variables like CC to empty string. This caused Makefiles ?= syntax to
skip sane fallbacks logic. This updates this to test for empty string
and performs assignment if the emtpy string is there.

BUG=chromium:788213
TEST=FEATURES=test emerge-reef mosys
FEATURES=test USE=static emerge-reef mosys
BRANCH=none

Change-Id: Ic6559c25681a5f5d37cfb79684f93573c68738a1
Reviewed-on: https://chromium-review.googlesource.com/788233
Commit-Ready: Jason Clinton <jclinton@chromium.org>
Tested-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>