mosys: cros_config C interface using structs
This is a cursory patch to get a feel on the approach.
If everyone is basically lgtm on this, I'll start code gen'ing the
cros_config_data.c file as part of the cros_config build and then stitch
it the rest of the way into mosys.

BUG=b:66950157
TEST=None
BRANCH=none

Change-Id: I4ac9619b04f56fe0f1b5cc8408ced09f2acba335
Reviewed-on: https://chromium-review.googlesource.com/998978
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>