cros_config_struct: Clarify config lookup stages
At present we have a fallback from FDT to yaml. Really this is not a great
idea, since if we are using yaml we would want to report errors. But
normally if we are using yaml there is no FDT config, so the effect should
be the same.

But it is useful to see what stage we are at and when we have given up on
finding a yaml config. Add some more debug statements for this.

BUG=chromium:833858
TEST=FEATURES=test emerge-grunt --nodeps mosys
Run mosys platform model -vvvv

Change-Id: I33b82e2bcf22cdab88372da96a4ac84c3ca93153
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1113901
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>