Revert "mosys: Add integration test"
This reverts commit d402d8411b3686d23f6f8149843a5cf2ad61f655.

By putting a printout in the kernel in
"arch/arm64/mm/mmap.c:devmem_is_allowed()" it can be seen that
"/dev/mem" is being accessed.  Reverting this change stops the bad
accesses.

It is believed that the problem is that the bad commit started calling
internal_read_identity_info() on ARM devices.  This calls
smbios_sysinfo_get_name().  That calls smbios_find_table().  If you
keep digging, you'll find that this eventually translates into reads
of address ranges in "/dev/mem".  It's not a good idea to do this.

This commit can be re-landed when it can be confirmed that random
/dev/mem accesses are no longer generated by it on ARM or ARM64
devices.

BUG=chromium:878145
TEST=mosys is no longer accessing /dev/mem

Fixes: d402d8411b36 ("mosys: Add integration test")
Change-Id: Iabb27a9dd5e2a8e651302a43373eea19cfb110f9
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1194284
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>