Update docs and build to reflect Rust frontend and tests enabled
Samantha enabled the Rust frontend and added Rust unit tests over the
summer. Update the docs to reflect that and make it clearer how to work
with Mosys locally, now.

This also removes the obsolete mosys.c and the meson.build rule since
those are effectively dead and have been for awhile in the ebuild.

BUG=none
TEST=copy-pasted new README contents and confirmed they work

Change-Id: I00b48d9bc682cffb9894cb9906f83c6d784f4880
Reviewed-on: https://chromium-review.googlesource.com/1195728
Commit-Ready: Jason Clinton <jclinton@chromium.org>
Tested-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>