mosys: check platform compatible on ARM/unibuild
Check that the ARM platform is compatible before trying to read the
unibuild config. This is failing on the new cheza platform and
identifying it as a gru because it didn't filter out at the platform
level. We read the /proc/device-tree/compatible file and see if the
'google,gru' identifier is in the compatible device-tree string.

We will need to check for 'google,cheza' in the new setup for the
cheza platform.

BUG=b:111891266
TEST=local builds on cheza and scarlets, tryjobs

Change-Id: Ie895a258849de227df65dabfd8b256cc0e513970
Reviewed-on: https://chromium-review.googlesource.com/1204695
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Gregory Meinke <gmeinke@chromium.org>
Reviewed-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>