New ARM specific function to read sku info
Broke out the cros_config_read_sku_info into an ARM specific function
to read the FDT information. Allow the calling devices to pass in an
array of compatible devices and simplify the naming to exclude the
leading 'google,' for the matching platform id.

BUG=b:114131525
TEST=local builds, test mosys on scarlet, tryjobs

Change-Id: Ia248c07b1ff2bb7a85a63373dd7878264ece5a97
Reviewed-on: https://chromium-review.googlesource.com/1217465
Commit-Ready: Gregory Meinke <gmeinke@chromium.org>
Tested-by: Gregory Meinke <gmeinke@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: C Shapiro <shapiroc@google.com>