sarien: Add empty wilco ec handling
Sarien was set to use the cros_ec callbacks, which will not work
on this platform.

As a placeholder provide a wilco_ec callback that returns hardcoded
values.  Eventually this will be able to read from sysfs but the
kernel code is not yet landed.

This is required for FAFT testing.

BUG=b:121380403
BRANCH=none
TEST=manual:
localhost$ mosys ec info
google | wilco | 0.0.0

Change-Id: Ifccf70191bc5fc90708900ea312e29b56c75e770
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1413931
Reviewed-by: Simon Glass <sjg@chromium.org>