Supply an updated SPD table for Grunt devices
Mosys is reading the SPD table from the RO firmware, instead of the
firmware that was actually used to boot.  On careena, the part number
for several of the memory chips was updated after locking the RO
firmware, causing mosys to not be able to find the SPD.

This is a temporary fix to make the SPD read for HWID calculation
work and should be updated when a better solution is in place.

BUG=b:122636013
TEST=Run on barla with samsung memory, get valid SPD data
mosys memory spd print all

Change-Id: I14127b6e6028f206fd7a743b89cb2d7dee7f3d70
Signed-off-by: Martin Roth <martinroth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1415835
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Tested-by: Cheng-Han Yang <chenghan@chromium.org>