sarien: Set SPD page to 0 before reading
In some cases FSP might leave the SPD pointer set to page 1,
which will cause mosys to not be able to read the SPD info.
To fix this force a write to address 0x36 on the same I2C bus
which will set the page back to 0.

This change also switches to use a hard coded DIMM count rather
than reading it from SMBIOS.  This is because if there is a
DIMM only installed in slot B it will not be found if SMBIOS
says there is only 1 DIMM in the system.

BUG=b:123211109
BRANCH=none
TEST=test 'memory spd print all' on sarien/arcada with both
DIMMs installed, and with one DIMM installed in slot A or B.

Change-Id: Idcb872a8a54e374c6c344613772a24af7ba3e55e
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1437739
Tested-by: Lijian Zhao <lijian.zhao@intel.com>
Tested-by: EricR Lai <ericr_lai@compal.corp-partner.google.com>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Lijian Zhao <lijian.zhao@intel.com>
Reviewed-by: Shelley Chen <shchen@chromium.org>