google/reef: Miniml hack to enable 'model' for electro and basking
The firmware updater needs to know which model current device is, and
the ToT mosys has been refactored with huge changes for that (model
based on SKU). Since ToT updaters stop embedding utility binaries (i.e.,
no mosys bundled) we have to backport this feature to factory branch.

For Electro and Basking, we need to identify from SKU ID; and this is
the minimal hack to enable that (for $(mosys platform model)).

BUG=b:130115628
TEST=emerge-reef mosys

Change-Id: I1617b5200fce29a6da5d4f8795668bd687ace7cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/1574862
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>