Drop 'eeprom csum' command
This command does not seem to be used in Chrome OS. I'm not sure if it is
useful on the command line or not. But it has not been updated in 8 years
and does not seem to be used in any tests.

According to dlaurie it dates back to Mario/Alex/ZGB before we had our own
firmware.

So drop it. With this we can drop some crypto code too.

BUG=chromium:980752
TEST=emerge-sarien sys-apps/mosys

Change-Id: I53b360330ff7507348228f43b640ae98e41910d6
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/1746502
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@google.com>