Make ec callback destruction common
There was a common pattern of writing a destroy function for each
platform intf calling each of the EC destroy methods if they were
defined. Simply make this code common so that we can reduce the amount
of code copy/pasted to each platform.

BUG=chromium:990438
TEST=sudo FEATURES=test emerge mosys

Change-Id: Ia470f45fa994714e6e3347e6d5430802843c14a6
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1827924
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>