rambi/slippy: Remove gpio callbacks and call driver directly
The memory module in rambi and slippy used the gpio callback interface
to the platform to call the underlying read functions, but this
function is constant in the interface.

Since I want to ultimately remove gpio callbacks from mosys now that
the command is gone, refactor this.

BUG=chromium:990438
TEST=sudo FEATURES=test emerge mosys

Change-Id: If2a4fa1c90ecd15cfd8aeb61f9ca0f0376667074
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1881398
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>