Add strlcpy
strlcpy is from BSD-land and provides a safe way to copy strings
within memory bounds. Unlike strncpy, strlcpy guarantees the string
will be terminated with a null byte if the buffer is not large enough,
and returns the number of bytes copied.

strlcpy will be used to clean up a lot of memory-related bugs in
follow up CLs.

BUG=chromium:1018847
TEST=strlcpy works as intended with usage in follow-up CLs

Change-Id: Ibf8cbafd3638e91ebabd0926b885c5c7d07ea614
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/1915173
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Will Bradley <wbbradley@chromium.org>