Stack allocate return string values from EC callbacks
Previously, these functions followed a paradigm to malloc their return
value, which leads to memory leaks as nobody ever freed the values, or
they tried to use add_destroy_callback which is broken.

Change to a safer paradigm of passing the return buffer and buffer size
as parameters to the functions.

BUG=chromium:1018847
TEST=emerge-drallion mosys, verify ec command returns correct
info (tests wilco EC)
TEST=emerge-kukui mosys, verify ec command returns correct info (tests
cros EC)

Change-Id: I41247fa110d9cb84097e855d0e372c0bfb9c7454
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/1915177
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Will Bradley <wbbradley@chromium.org>