Add read_file helper
This is used in follow-up CLs to simplify repeated logic of opening
and reading files. Implement this as common, as the other places it
was used memory leaks were frequent, and by making callers pass in a
buffer for us to put data into, it'll help future callers choose an
appropriate way to allocate.

BUG=chromium:1018847
TEST=in follow up CLs, the function works as expected to read FDT and
ACPI info

Change-Id: I2292cd2b52e5cef0cfe95b494ee885fe468f1d59
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/1915176
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>