Remove Android support
Support was added to allow Mosys to run on Android for the Ryu project
years ago. The project was cancelled and this code is now going
unused.

Goal was to strip out Android support as cleanly as possible without
making large refactors. Flashrom support is pretty disgusting and
could use a refactor, but that will be handled separately (or may not
be necessary at all if we can drop eeprom support in a reasonable
schedule). So hence I didn't really care about cleaning up the mess of
strdups, etc.

BUG=chromium:990438
TEST='mosys eeprom map' still functions on samus

Change-Id: I70816fb0c96e6a960cacd7dab6e2c93000ac2701
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/1949237
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>