Remove nvram clear/list/dump, deprecate nvram vboot
We've cleared up all references in Chrome OS to nvram clear, list, and
dump. The remaining usage of the nvram command is by crossystem for
vboot parameters, which is in the process of being migrated to
crossystem (crbug.com/1032351, crbug.com/1030473). Make the help
message clear we don't want anyone else to start using it in the mean
time.

BUG=chromium:1030473
TEST=compiles

Change-Id: I9efc5c5db690e1623a767acade2a5693f39b5005
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/1968097
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>