cros_ec: poll cros_ec chardev until available during initialization
Apparently removing the "big lock" in mosys broke some
things (b/154444676). nyan_kitty stores vbnv context in the
EC (deprecated for new devices), and crossystem reads this during
startup (via session_manager) thru mosys.

What is currently happening is that opening /dev/cros_ec will fail
with EACCES, EBUSY, or EAGAIN, indicating the chardev is open by
either the kernel or another process.

Historically, this never happened as the broad
/run/lock/firmware_utility_lock was in the way (a convention-based
userspace lock for pretty much anything). But we got rid of that in
favor of locking individual resources instead.

(Yeah, kinda sucks we gotta use a sleep-poll pattern, but Linux has
no way to do a blocking open, or the equivalent of select(2) for an
open call.)

Additionally, this commit refactors some of the cros_ec code so that
we don't have to initialize the EC unless we are using it. This
required going and deleting the cros_ec_setup from the setup_post of
many platforms, hence the noise in the platform directory.

BUG=b:154444676
TEST=boot nyan_kitty, /home/chronos/crash does not get a truckload of
     dumps from mosys. Boot speed regression has gone away.

Change-Id: I5e249738849d7f19dbdb6e96a56ea62e9fa8c727
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/2158106
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>