platform/sarien: use SMBIOS type 17 memory info for Sarien & Drallion
The firmware for sarien, arcada, and drallion all expose memory
information thru SMBIOS type 17 tables, which is the preferred way to
expose memory information to the OS.  Switch to using smbios_memory_cb
for both platforms.

In particular, this is desired because Deltaur & Deltan were planning
on using a single system image, but two different mosys platforms for
the memory support.  This does not align with the goal of making mosys
compile for a single platform on a single board, and needs fixed.
After this change, Drallion, Deltan, and Deltaur can all unify on the
Sarien mosys platform, and the Drallion mosys platform can be removed
once we update the model.yaml files.

BUG=chromium:1070692,b:151724803
TEST="mosys memory spd print all" on sarien, arcada, and drallion

Change-Id: I137416f0537aeaf73e9ced10bfbae8cc64892022
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/2169844
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-by: Mathew King <mathewk@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>