Plumb thru meson option for single platform support
This adds a meson option, "platform_intf", that when set to a
non-empty string, will define CONFIG_SINGLE_PLATFORM to the name of
the platform supported (as a string).

Actually doing the compilation of the right platform will involve a
little bit of macro-trickery and changing the platform selection code,
so that will be done in some follow-up CLs.

BUG=chromium:1070692
TEST="meson out -Dplatform_intf=Foo" generates out/config.h with:
     #define CONFIG_SINGLE_PLATFORM "Foo"

Change-Id: I3860f6487fa9bf40f99151a4c7af1c72cd585764
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/2171763
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>