Resurrect the C main function
The mosys rewrite in Rust was nowhere close to completion, and the
people who were leading this effort are not on the project anymore.

This ressurects the main function, as it existed in 2017 (look in the
git history for /mosys.c at the root of the repository).  The only
changes that were made were:

- Remove the big lock, as mosys has removed this feature.
- Minor style fixes.

The main function is still plumbed thru a tiny bit of rust code which
handles the minijail and seccomp setup.  The plan is to remove all
Rust code, so we need to make a decision if mosys actually requires
minijail, or if this is something which was unnecessary and can
be dumped.  If we want to keep it, we can use libminijail0, a C
library, instead of the Rust io_jail library in this repo.
Regardless, this should be solved in a separate CL to keep this one
small.

BUG=chromium:1024427,chromium:992520
TEST=compile, test mosys manually, all commands and flags are working

Change-Id: I1b2af9532e0b701ff6a9801a039da1f9739352b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/2176190
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>