lib/smbios: use read_file for smbios_scan_sysfs
I couldn't get this scanft version of smbios_scan_sysfs to work.  This
is one of the few spots that scanft is used in mosys, and it's way
overkill when we can just use read_file.

Presumably, this code was never getting used as we were always going
thru /dev/mem to read this system information.  We want to change that
to use sysfs when we can.

BUG=chromium:1078420
TEST=probe_smbios is working with follow-up CL applied

Change-Id: Ifd23d1b3cd2e984e855d6e5066b71008235b9342
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/2182983
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>