Port src/main.rs to C
The only remaining bits of Rust code left in this repo are to handle
the minijail and seccomp functionality.  This was accomplished using
the io_jail wrapper for the minijail library, which very closely
mirrors the API of the C minijail library.

This commit ports the remaining Rust functionalities to C.  That is,
mains/mosys.c was designed to look as close to src/main.rs as
possible, but without the Rust.

The ebuild is still using the Rust build functionality and main
function.  A follow-up CL will "flip the switch" on the ebuild to
start using this C main function, and then we may remove all remaining
Rust code from this repository.

BUG=chromium:1024427,chromium:992520
TEST=meson out && ninja -C out, manually test out/mains/mosys

Change-Id: I20547844eb0734e9e9854a27f656ea4a29fed297
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/2198393
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>