cros_ec: remove retry-logic to initialize device
This was introduced so that devices using the EC to store vboot
nvstorage would provide correct values on bootup (when the EC is not
ready yet) instead of crossystem failing.  However, this introduced a
boot-time regression on nyan_kitty and some veyron_* devices.

They are almost AUE, so just revert to the old behavior.

BUG=chromium:1091903
TEST=monitor crosbolt

Change-Id: I478923da5ac988de5f27d2be51ec91c73577d32b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/2246294
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>