platform/arm/oak: drop "mosys pd" command support
pd command was only available on the Oak reference board, and there's
some special code here to handle deleting the command from the
platform when a non-reference board is probed.

Since we are moving to unibuild, we have to get rid of the special
probing code.  PD command was not that important, and the reference
board should be considered unsupported anyway now that we launched
real devices under this platform.

Drop Rowan as the project was canceled.

BUG=chromium:1115620
TEST=no changes to hana mosys behavior

Change-Id: Ic04fed36830ced480151cff3507ca761be81eeae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/mosys/+/2353239
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>